bump up a version bound for ghc 7.8
this doesn't totally make it work by itself, a few more hacks
in other packages are required. property-list has an unsafePerformIO
that needs to be reimported, and a package it depends on
(recursion-schemes) needs substantial fixes to Typeable/Data stuff
which was changed a lot in 7.8. for now, if you just remove the
Typeable/Data declarations entirely jammittools still works.